Edited table + improved query

+++ sendmessage/app.js
@@ -8,10 +8,19 @@ const ddb = new AWS.DynamoDB.DocumentClient({ apiVersion: '2012-08-10', region: 
 const { TABLE_NAME } = process.env;
 
 exports.handler = async event => {
-  let connectionData;
+  let openConnections;
+
+  console.log('Incoming event', event);
   
   try {
-    connectionData = await ddb.scan({ TableName: TABLE_NAME, ProjectionExpression: 'connectionId' }).promise();
+    openConnections = await ddb.query({
+      TableName: TABLE_NAME,
+      IndexName : "openConnectionsIndex",
+      KeyConditionExpression: "isOpen = :openValue",
+      ExpressionAttributeValues: {
+        ":openValue": "true"
+      },
+    }).promise();
   } catch (e) {
     return { statusCode: 500, body: e.stack };
   }
@@ -20,24 +29,33 @@ exports.handler = async event => {
     apiVersion: '2018-11-29',
     endpoint: event.requestContext.domainName + '/' + event.requestContext.stage
   });
-  
-  const postData = JSON.parse(event.body).data;
-  
-  const postCalls = connectionData.Items.map(async ({ connectionId }) => {
+
+  const sender = event.requestContext.connectionId;
+  const messageBody = JSON.parse(event.body);
+  const postData = {
+    data: messageBody.data,
+    sender,
+  };
+
+  const postCalls = openConnections.Items.map(({ connectionId }) => {
+    if (connectionId === sender) {
+      return;
+    }
+
     try {
-      await apigwManagementApi.postToConnection({ ConnectionId: connectionId, Data: postData }).promise();
+      return apigwManagementApi.postToConnection({ ConnectionId: connectionId, Data: JSON.stringify(postData) }).promise();
     } catch (e) {
       if (e.statusCode === 410) {
         console.log(`Found stale connection, deleting ${connectionId}`);
-        await ddb.delete({ TableName: TABLE_NAME, Key: { connectionId } }).promise();
-      } else {
-        throw e;
+        return ddb.delete({ TableName: TABLE_NAME, Key: { connectionId } }).promise();
       }
+      console.log(`Unexpected error while sending the message`, e);
+      throw e;
     }
   });
   
   try {
-    await Promise.all(postCalls);
+    await Promise.allSettled(postCalls);
   } catch (e) {
     return { statusCode: 500, body: e.stack };
   }
